Moving test _helpers back into library.
This is preparation for subpackage split-up.
